unify field conversions
